refactor(db): move feed store ownership
